+++ addie/configuration/import_configuration.py
@@ -6,8 +6,8 @@ from addie.configuration.config_file_name_handler import ConfigFileNameHandler
 from addie.utilities.conversion import str2bool
 from addie.utilities.gui_handler import GuiHandler
 from addie.autoNOM.step1_gui_handler import Step1GuiHandler
-from addie.step2_handler.import_table import ImportTable
-from addie.step2_handler.step2_gui_handler import Step2GuiHandler
+from addie.processing_idl.import_table import ImportTable
+from addie.processing_idl.step2_gui_handler import Step2GuiHandler
 
 
 class ImportConfiguration(object):
