+++ addie/processing_idl/create_sample_files.py
@@ -1,7 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 import os
 
-from addie.step2_handler.table_handler import TableHandler
+from addie.processing_idl.table_handler import TableHandler
 
 
 class CreateSampleFiles(object):
