+++ addie/processing_idl/run_sum_scans.py
@@ -1,7 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 import os
 from qtpy.QtCore import Qt
-from addie.step2_handler.step2_gui_handler import Step2GuiHandler
+from addie.processing_idl.step2_gui_handler import Step2GuiHandler
 
 
 class RunSumScans(object):
