+++ scripts/addie
@@ -53,19 +53,19 @@ from addie.initialization.init_step1 import InitStep1
 
 from addie.autoNOM.step1_gui_handler import Step1GuiHandler
 from addie.autoNOM.run_step1 import RunStep1
-from addie.step2_handler.load_table_intermediate_step_interface import loadTableIntermediateStepInterface
-from addie.step2_handler.undo_handler import UndoHandler
+from addie.processing_idl.load_table_intermediate_step_interface import loadTableIntermediateStepInterface
+from addie.processing_idl.undo_handler import UndoHandler
 
 from addie.initialization.init_step2 import InitStep2
-from addie.step2_handler.populate_master_table import PopulateMasterTable
-from addie.step2_handler.populate_background_widgets import PopulateBackgroundWidgets
-from addie.step2_handler.step2_gui_handler import Step2GuiHandler
-from addie.step2_handler.table_handler import TableHandler
-from addie.step2_handler.create_sample_files import CreateSampleFiles
-from addie.step2_handler.create_ndsum_file import CreateNdsumFile
-from addie.step2_handler.run_ndabs import RunNDabs
-from addie.step2_handler.run_sum_scans import RunSumScans
-from addie.step2_handler.run_thread import RunThread
+from addie.processing_idl.populate_master_table import PopulateMasterTable
+from addie.processing_idl.populate_background_widgets import PopulateBackgroundWidgets
+from addie.processing_idl.step2_gui_handler import Step2GuiHandler
+from addie.processing_idl.table_handler import TableHandler
+from addie.processing_idl.create_sample_files import CreateSampleFiles
+from addie.processing_idl.create_ndsum_file import CreateNdsumFile
+from addie.processing_idl.run_ndabs import RunNDabs
+from addie.processing_idl.run_sum_scans import RunSumScans
+from addie.processing_idl.run_thread import RunThread
 
 from addie.step3_handler.step3_gui_handler import Step3GuiHandler
